+++ packages/adapter-oas/src/resolvers.ts
@@ -2,8 +2,7 @@ import { pascalCase } from '@internals/utils'
 import type { ast } from '@kubb/ast'
 import { Diagnostics } from '@kubb/core'
 import { formatMap, SCHEMA_REF_PREFIX, specialCasedFormats, structuralKeys } from './constants.ts'
-import { isReference } from './guards.ts'
-import { isJsonMimeType } from './mime.ts'
+import { isJsonMimeType, isReference } from './oas.ts'
 import { getRequestContent, getResponseByStatusCode } from './operation.ts'
 import { dereferenceWithRef, resolveRef } from './refs.ts'
 import type { ContentType, Document, MediaTypeObject, Operation, ParameterObject, ResponseObject, SchemaObject, ServerObject, ServerOptions } from './types.ts'
